+++ lbrynet/core/Wallet.py
@@ -941,6 +941,7 @@ class Wallet(object):
                                             bid, certificate_id, claim_address, change_address)
 
         if not claim['success']:
+            log.error(claim)
             msg = 'Claim to name {} failed: {}'.format(name, claim['reason'])
             raise Exception(msg)
 
@@ -1312,9 +1313,14 @@ class LBRYumWallet(Wallet):
         return defer.succeed(True)
 
     def _check_large_wallet(self):
-        if len(self.wallet.addresses(include_change=False)) > 1000:
-            log.warning(("Your wallet is excessively large, please follow instructions here: ",
-                         "https://github.com/lbryio/lbry/issues/437 to reduce your wallet size"))
+        addr_count = len(self.wallet.addresses(include_change=False))
+        if addr_count > 1000:
+            log.warning("Your wallet is excessively large (%i addresses), "
+                        "please follow instructions here: "
+                        "https://github.com/lbryio/lbry/issues/437 to reduce your wallet size",
+                        addr_count)
+        else:
+            log.info("Wallet has %i addresses", addr_count)
 
     def _load_blockchain(self):
         blockchain_caught_d = defer.Deferred()