+++ lbrynet/daemon/Daemon.py
@@ -514,7 +514,7 @@ class Daemon(AuthJSONRPCServer):
 
     @defer.inlineCallbacks
     def _setup_lbry_file_manager(self):
-        log.info('Starting to setup up file manager')
+        log.info('Starting the file manager')
         self.startup_status = STARTUP_STAGES[3]
         self.stream_info_manager = DBEncryptedFileMetadataManager(self.db_dir)
         self.lbry_file_manager = EncryptedFileManager(
@@ -670,8 +670,7 @@ class Daemon(AuthJSONRPCServer):
             self.streams[sd_hash] = GetStream(self.sd_identifier, self.session,
                                               self.exchange_rate_manager, self.max_key_fee,
                                               self.disable_max_key_fee,
-                                              conf.settings['data_rate'], timeout,
-                                              file_name)
+                                              conf.settings['data_rate'], timeout)
             try:
                 lbry_file, finished_deferred = yield self.streams[sd_hash].start(claim_dict, name)
                 yield self.stream_info_manager.save_outpoint_to_file(lbry_file.rowid, txid, nout)
@@ -917,7 +916,7 @@ class Daemon(AuthJSONRPCServer):
         defer.returnValue(lbry_file)
 
     @defer.inlineCallbacks
-    def _get_lbry_files(self, return_json=False, full_status=False, **kwargs):
+    def _get_lbry_files(self, return_json=False, full_status=True, **kwargs):
         lbry_files = list(self.lbry_file_manager.lbry_files)
         if kwargs:
             for search_type, value in iter_lbry_file_search_values(kwargs):
@@ -2041,6 +2040,8 @@ class Daemon(AuthJSONRPCServer):
             'claim_address': claim_address,
             'change_address': change_address,
             'claim_dict': claim_dict,
+            'channel_id': channel_id,
+            'channel_name': channel_name
         })
 
         if channel_id: